% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackout_cells.R
\name{blackout_cells}
\alias{blackout_cells}
\title{Preparing a data frame of zeros to cover the background for visualization in heatmap3}
\usage{
blackout_cells(truncated_m)
}
\arguments{
\item{truncated_m}{A final, normalized interaction matrix generated by truncate_matrix().}
}
\value{
A dataframe of zeros to cover background for visualization.
}
\description{
Preparing a data frame of zeros to cover the background for visualization in heatmap3
}
