% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_matrix.R
\name{inter_matrix}
\alias{inter_matrix}
\title{Generating a basic contact matrix from HOMER tags file}
\usage{
inter_matrix(bgmodel)
}
\arguments{
\item{bgmodel}{A HOMER tags file read in by read_homer().}
}
\value{
A basic contact (interaction) matrix.
}
\description{
Generating a basic contact matrix from HOMER tags file
}
\examples{
-
}
