% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_param.R
\name{norm_param}
\alias{norm_param}
\title{Generating a matrix of the sum of the reads mapping to both fragments involved in the interaction}
\usage{
norm_param(bins)
}
\arguments{
\item{bins}{a matrix of restriction fragment sized bins generated by make_bins().}
}
\value{
A matrix of the sum of the reads for normalization.
}
\description{
Generating a matrix of the sum of the reads mapping to both fragments involved in the interaction
}
\examples{
-
}
