% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalizing a binned contact matrix by dividing it by the sum of the reads mapping to both fragments involved in the interaction}
\usage{
normalization(bins, norm_param)
}
\arguments{
\item{bins}{a matrix of restriction fragment sized bins generated by make_bins().}

\item{norm_param}{A binned raw matrix generated by norm_param().}
}
\value{
A normalized contact matrixs.
}
\description{
Normalizing a binned contact matrix by dividing it by the sum of the reads mapping to both fragments involved in the interaction
}
\examples{
-
}
