% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plots a heatmap of a normalized interaction matrix}
\usage{
plot_results(
  truncated_m,
  black_cells,
  title = "T2C results",
  colors_palette = c("blue", "deepskyblue2", "green", "yellow", "orange", "red", "red3")
)
}
\arguments{
\item{truncated_m}{a binned contact-matrix generated by \code{truncate_matrix} function.}

\item{black_cells}{A data frame of array indices generated by \code{blackout_cells} function.}

\item{title}{A plot title, default = T2C results.}

\item{colors_palette}{colors passed to \code{colorRampPalette} function.}
}
\value{
A heatmap3 plot of contacts.
}
\description{
Plots a heatmap of a normalized interaction matrix
}
