% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_matrix.R
\name{truncate_matrix}
\alias{truncate_matrix}
\title{Limit extreme matrix values to a fraction of the maximum value of input matrix to improve visualization}
\usage{
truncate_matrix(ext_matrix, limit = 0.25)
}
\arguments{
\item{ext_matrix}{A matrix generated by extending_matrix().}

\item{limit}{A coefficient by which the maximum values is obtained - to be determined by trial and error.}
}
\value{
A finished, extended interaction matrix which is used as an input to blackout_cells() and plot_results().
}
\description{
Limit extreme matrix values to a fraction of the maximum value of input matrix to improve visualization
}
\examples{
\dontrun{
truncated_m <- truncate_matrix(extended_matrix, limit = 0.25)
}
}
