% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{Generating restriction fragment sized bins}
\usage{
make_bins(regions, inter_matrix)
}
\arguments{
\item{regions}{Regions between restriction enzyme cutting sites.}

\item{inter_matrix}{A basic contact matrix generated by \code{inter_matrix} function.}
}
\value{
A matrix of restriction fragment sized bins of class \code{rawbins}.
}
\description{
Generating restriction fragment sized bins
}
