% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalizing a binned contact matrix by dividing it by the sum of the reads in bins (mapping to both fragments involved in the interaction)}
\usage{
normalization(bins, norm_param)
}
\arguments{
\item{bins}{a matrix of restriction fragment sized bins generated by \code{make_bins} function.}

\item{norm_param}{A binned raw matrix generated by \code{norm_param} function.}
}
\value{
A normalized contact matrix of class \code{normalized}.
}
\description{
Normalizing a binned contact matrix by dividing it by the sum of the reads in bins (mapping to both fragments involved in the interaction)
}
