% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plots a heatmap of a contact-matrix}
\usage{
plot_results(
  obcieta,
  title = c("T2C results"),
  black_cells,
  colors_palette = c("blue", "deepskyblue2", "green", "yellow", "orange", "red", "red3")
)
}
\arguments{
\item{obcieta}{a binned contact-matrix generated by truncate_matrix().}

\item{title}{A plot title, default = T2C results.}

\item{black_cells}{A data frame of array indiced generated by blackout_cells().}
}
\value{
A heatmap3 plot of contacts.
}
\description{
Plots a heatmap of a contact-matrix
}
\examples{
-
}
