% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_homer.R
\name{read_homer}
\alias{read_homer}
\title{Reads in HOMER tags file, creates a data frame containing reads in a region of interest}
\usage{
read_homer(fn, coord)
}
\arguments{
\item{fn}{path to file and filename of tags generated by HOMER.}

\item{coord}{A vector with three elements c(chrom, start, end) like c(7, 110901160, 111101971).}
}
\value{
A cleaned-up data frame of HOMER tags in a region of interest.
}
\description{
Reads in HOMER tags file, creates a data frame containing reads in a region of interest
}
