% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_matrix.R
\name{truncate_matrix}
\alias{truncate_matrix}
\title{Limit matrix values to the chosen value}
\usage{
truncate_matrix(ext_matrix, limit = 0.25)
}
\arguments{
\item{ext_matrix}{A matrix generated by extending_matrix().}

\item{limit}{A number to which limit the matrix to.}
}
\value{
A finished, axtended contact matrix which is used as an input to blackout_cells() and plot_results().
}
\description{
Limit matrix values to the chosen value
}
\examples{
obcieta <- truncate_matrix(extended_matrix, limit = 0.2)
}
